mgsrc/backend/requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.9
python-socketio[client]>=5.11.0
Flask-SocketIO>=5.3.6
pydantic>=2.6.0
python-dotenv>=1.0.1
